requirements.txt
# Flask Web Framework and essentials
Flask>=2.0.0
Werkzeug>=2.0.0

# Google Cloud Libraries
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0       # For optional LLM filename suggestion

# Audio Processing
pydub>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Google API Core dependencies (often needed, good to include explicitly)
google-api-core>=2.0.0
googleapis-common-protos>=1.50.0
google-auth>=2.0.0
elevenlabs>=0.3.0
